tests: cleanup dupe mock hive blocks d
